+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added
+- Ability to change wallpaper as part of the ransomware simulation on Windows.
+  #1247
 
 ### Changed